automatically backed up by learn

+++ lib/create.sql
@@ -1,7 +1,7 @@
 CREATE TABLE projects (id INTEGER PRIMARY KEY,
                        title TEXT,
                        category TEXT,
-                       funding_goal FLOAT,
+                       funding_goal DECIMAL,
                        start_date DATE,
                        end_date DATE);
